# riscv: timer: driver revamp