@teburd 2023-02-17
The timer code simplification and test results speak pretty well to this! I do wonder if the reduced maximum compare diff might be modestly detrimental when long duration sleeps are desired, but probably not a big loss here.