@npitre 2023-02-17
> I do wonder if the reduced maximum compare diff might be modestly 
> detrimental when long duration sleeps are desired, but probably not 
> a big loss here.

On RV64, the restriction based on unsigned long amounts to 64 bits so
no change from before there. The limiting factor here is the int32_t type
for tick reporting. With the default CONFIG_SYS_CLOCK_TICKS_PER_SEC=10000
this means the maximum sleep duration is 214748 seconds, or half of that
if considering one bit of room for IRQ servicing latency (of course IRQ
latency will never be that large but that makes the code simpler).
That's about 29.8 hours. Longer delays will see intermediate wake-ups
every 29.8 hours.

On RV32 it depends on the HW clock cycle rate and that varies wildly.
The hardware I have here at the moment uses 60MHz. Given an unsigned long
is 32 bits, that gives us:

```
    CYC_PER_TICK = 60000000 / 10000 = 6000
    max_ticks = 2^32 / 2 / 6000 = 357913
```

That's 35 seconds max without intermediate wake-ups.

However, if sleep time is important, then you shouldn't need a 10kHz
tick rate. Going with 100 might still be sufficiently precise while
providing almost an hour of uninterrupted sleep. And if sleep time is
really important then you'd better select a much slower hw timer clock
rate in the first place anyway for power consumption reasons.

And if that still isn't satisfactory for your application then you just
need to redefine `cycle_diff_t` to an `uint64_t` in order to benefit from
the same range and accuracy as on RV64, albeit with a 64-bit division
runtime cost.