@andyross 2023-02-21
> - The MIN_DELAY thing is pointless. If the delay gets close or even behind current time then the IRQ will be triggered right away. This is unlikely to happen very often anyway so the constant overhead is uncalled for.

Are we sure about that?  I have a vague memory that this was a fix.  It's true that the machine timer docs **specify** the interrupt latch as a true comparator (e.g. the interrupt signal is equal to the sign bit of "comparator_time - current_time"), and I have a very vague sense that that's how it was written originally, but then had to be reverted because it turned out that some device or another didn't do it right and implemented an equality test like most other hardware?

Worth digging through the git logs to verify.

FWIW: a really nice project would be to unify the timer driver boilerplate.  With only a handful of exceptions, all the hardware works as free-running up counters (of varying bit lengths) with an expiration time register, and just a little bit of tuning (like MIN_DELAY for "how soon in the future can we set an interrupt reliably").  We tend to fix the same bugs repeatedly and return the same notions.