@npitre 2023-02-21
> Are we sure about that?  I have a vague memory that this was a fix.
> It's true that the machine timer docs **specify** the interrupt latch
> as a true comparator (e.g. the interrupt signal is equal to the sign
> bit of "comparator_time - current_time"), and I have a very vague sense
> that that's how it was written originally, but then had to be reverted
> because it turned out that some device or another didn't do it right
> and implemented an equality test like most other hardware?

So far, I have proof that this is a signed comparator and not an equality
comparator on both QEMU and an actual hardware implementation (just tested
it with explicit behind values).

> Worth digging through the git logs to verify.

The git history doesn't mention anything about it in the commit where this
`MIN_DELAY` was introduced (by yourself by the way  ;-)) although that
commit mentions fixing an unrelated bug.

Yet, this `MIN_DELAY` caused issues later on and had to be reduced in a
subsequent commit. I personally hate those magic values that are never
right.

So, given the lack of evidence this is needed, and given we're early in this
release cycle, I'd suggest this be merged as is and see if someone
complains... though this might be hard to trigger.

In any case, even if this was a simple equality comparator, having a magic
min value is not the way to go. Instead, it should be done as:

```
    set_clock_match_reg(cycle_val);
    if (unlikely((read_clock_reg() - cycle_val) >= 0)) {
        uint32_t ahead = 1;
        do {
            cycle_val = read_clock_reg() + ahead;
            ahead *= 2;
            set_clock_match_reg(cycle_val);
        } while ((read_clock_reg() - cycle_val) >= 0);
    }
```

The comparison with equality is needed as the comparator is mostly latched
on clock edge. This is sure to get ahead regardless of the clock rate vs
CPU speed.

In fact defensive programming would mean I should probably implement the
above regardless, possibly with a Kconfig option so only those who need it
will pay the overhead, and a simple test in the init function to warn
people when they actually need to turn it on. But in a separate PR maybe?