@npitre 2023-02-21
> FWIW: the test code around the set isn't sufficient.  Hardware itself has
> queuing/buffering of I/O that makes the race condition inherent.

Yeah, I know. I first met this issue with the Intel SA1100 timer, for
those who might still remember that chip.

It is not uncommon for writes to be posted even to IO area, however a
following read to the same IO area will be ordered, meaning that the read
will take place only after the preceding write has completed. That covers
the illustrated case here.