@npitre 2023-02-21
> I quite like this try loop to set the compare in the unlikely case
> there's a lot of unexpected/unlikely bus latency to set a new one up.

If the comparator is a simple equality comparator then this is absolutely
needed for reliability. But if the implementation is spec compliant then
the comparator should be a numerical one meaning the IRQ will be raised
when the match value is equal _or_ behind the clock value, in which case
there is no need for extra care in software.

Whether or not all implementations are spec compliant is another matter.